fix ResAddrMmap error



# message auto-generated for no-merge-commit merge:
!1948 merge master into master

fix ResAddrMmap error

Created-by: xz00xz
Commit-by: xz00xz
Merged-by: cann-robot
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
修复jetty va align down不符合预期的问题，需要将4096U改成0x1000
## 变更类型
请选择本次引入的变更类型：
<!--  [x] 表示选中 -->
- [x] 🐛 Bug修复
- [ ] ✨ 新特性
- [ ] 🚀 性能优化
- [ ] 📝 文档更新
- [ ] 📋 其他，请描述：

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。-->
<!-- 如果这个PR不涉及Issue，可填写"NA"。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于构造对应xx测试用例、二级冒烟、算子泛化等。-->
已完成的测试用例和场景：
1. 
2. 

补充的UT用例：

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 合入检查
<!-- 在正式合入前，请做好必要的代码测试，用例补充，软件代码风格检查等。提高合入效率。-->
<!--  [x] 表示选中 -->
- [x] 🧐 已经详细阅读了贡献指南（CONTRIBUTING.md），并遵守了其中的所有规定，包括但不限于commit message的格式、无效commit的合并等
- [x] 🔍 邀请 committer评论`/lgtm`前的必要检查
    - [ ] 🏷️ 标题中使用了合适的类型标签（如：`[feat]`, `[fix]`）
    - [ ] 📄 代码修改内容已简要描述，相关文档已更新
    - [ ] 📝 代码注释已更新，代码遵循项目整体代码风格
    - [ ] 🧪 代码UT测试已更新，覆盖率已达标
    - [ ] 🔬 验证方法已更新到"测试"部分
    - [ ] 🛠️ 代码已通过静态分析工具检查，无错误
    - [ ] 👥 代码检视/code review/同行评议和必要的代码串讲，确保代码质量
    - [ ] ✅ 代码检视意见已处理或答复，无未处理的检视意见
- [ ] 🚀 预约 前冒烟 用例前的必要检查
    - [ ] ✔️ 代码已有committer的`/lgtm` 和 模块committer的`/lgtm`评论
    - [ ] 🔧 代码已通过compile，编译无错误，无告警
    - [ ] 🖥️ 代码已通过基本功能本地测试或者在线测试，确保基本功能正常
- [ ] 🎯 预约 approver评论`/approve`，正式合入前的必要检查
    - [ ] 📊 前冒烟 用例已全量通过
    - [ ] 📦 新增功能已同步补充基本功能测试用例到前冒烟里

See merge request: cann/hcomm!1948

+++ src/platform/hccp/rdma_service/ctx/rs_ping_urma.c
@@ -317,7 +317,7 @@ STATIC int RsPingCommonInitSegCb(struct rs_cb *rscb, struct RsPingCtxCb *pingCb,
     CHK_PRT_RETURN(ret != 0, hccp_err("pthread_mutex_init seg_cb mutex failed, ret:%d", ret), ret);
 
     flag = ((unsigned long)pingCb->logicDevid << BUFF_FLAGS_DEVID_OFFSET) | BUFF_SP_SVM;
-    ret = DlHalBuffAllocAlignEx(segCb->len, RA_RS_4K_PAGE_SIZE,
+    ret = DlHalBuffAllocAlignEx(segCb->len, (unsigned int)RA_RS_4K_PAGE_SIZE,
         flag, (int)rscb->grpId, (void **)&segCb->addr);
     if (ret != 0) {
         hccp_err("DlHalBuffAllocAlignEx failed, length:0x%llx, dev_id:0x%x, flag:0x%lx, grp_id:%u, ret:%d",
